// webapp/src/app/core/api/v1/model/attachment.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Attachment { 
    attachmentId?: number | null;
    memberId?: number;
    categoryId?: number;
    categoryTitle?: string | null;
    policyPeriod?: number;
    createdBy?: string | null;
    originalFileName?: string | null;
    fileName?: string | null;
    fileDescription?: string | null;
    createdDate?: string;
    status?: string | null;
    applicationUser?: string | null;
    modifiedByDate?: string;
    modifiedBy?: string | null;
    applicationId?: number;
    rowId?: number;
    elementId?: number;
}